add ndarray converter (external blocks broken)
