+++ asdf/__init__.py
@@ -27,12 +27,24 @@ from ._version import version as __version__
 from .asdf import AsdfFile
 from .asdf import open_asdf as open
 from .config import config_context, get_config
-from .stream import Stream
-from .tags.core import IntegerType
+from .tags.core import IntegerType, Stream
 from .tags.core.external_reference import ExternalArrayReference
 
 
 def __getattr__(name):
+    if name == "stream":
+        import warnings
+
+        import asdf.tags.core.stream
+        from asdf.exceptions import AsdfDeprecationWarning
+
+        warnings.warn(
+            "asdf.stream is deprecated. Please use asdf.tags.core.stream",
+            AsdfDeprecationWarning,
+        )
+
+        return asdf.tags.core.stream
+
     if name == "AsdfExtension":
         # defer import to only issue deprecation warning when
         # asdf.AsdfExtension is used
